move substitutions parsing to new file
