+++ app/lib/client/fetcher.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'package:flutter/cupertino.dart';
 import 'package:internet_connection_checker/internet_connection_checker.dart';
 import 'package:intl/intl.dart';
 import 'package:rxdart/rxdart.dart';
@@ -78,7 +79,9 @@ class SubstitutionsFetcher extends Fetcher {
 
   @override
   Future<dynamic> _get() async {
-    final substitutionPlan = await client.getFullVplan();
+    debugPrint("Fetching substitution plan");
+    final substitutionPlan = await client.substitutions.getAllSubstitutions();
+    debugPrint("Fetched substitution plan");
 
     final Map filteredSubstitutionPlan = {"length": 0, "days": []};
