+++ app/lib/shared/shared_functions.dart
@@ -1,3 +1,2 @@
 void recordError(error,StackTrace stack) {
-
 }
